refactor(setup): extract launcher view to views/setup/launcher.py
PR 1 added ~210 LOC of direct-entry commands + the repost-launcher
button to ``cogs/setup_cog.py``, pushing it to 889 LOC and over the
S4.6 800-LOC fail threshold (`test_cog_size_under_fail_threshold`).

The fix follows the F-3 convention: a `_cog.py` file holds the
Discord-facing surface (the cog class, commands, listeners) and
helpers live in a sibling module. The launcher view fits naturally
under ``views/setup/`` alongside the rest of the setup panels.

Moves:
- ``_LAUNCHER_TITLE``, ``_LAUNCHER_DESC``, ``_START_LABELS_BY_STATUS``
- ``_build_launcher_embed``
- ``SetupLauncherView`` (all 8 buttons including the new Repost
  launcher button)
- ``_bot_can_send_in``, ``pick_launcher_channel``, ``post_launcher``

into ``disbot/views/setup/launcher.py``.

Keeps in ``cogs/setup_cog.py``:
- ``SetupCog`` with on_guild_join / on_ready listeners + the new
  ``!setup`` and ``/setup`` commands
- ``_resolve_hub_entry`` (cog-specific gating helper)
- Re-exports of ``SetupLauncherView``, ``pick_launcher_channel``,
  ``post_launcher`` so existing imports
  (``from cogs.setup_cog import SetupLauncherView``) keep working

Line counts after extraction:
- ``disbot/cogs/setup_cog.py``: 338 LOC (was 889)
- ``disbot/views/setup/launcher.py``: 550 LOC (new)

Test patch targets updated for the three repost-launcher tests:
``cogs.setup_cog.post_launcher`` → ``views.setup.launcher.post_launcher``
(``post_launcher`` is now a local binding inside the launcher
module's view callback). ``setup_session`` patches keep working via
the shared module object.

Also adds a small UX hint: the launcher embed description now
mentions the new ``!setup`` / ``/setup`` direct entry.

Verified: 3569 tests pass, ruff/black/isort/mypy clean.
